Refactoring: Memindahkan repository PendaftaranKonsultasi ke foldernya sendiri
