+++ src/main/java/Kelompok2_RPL/AplikasiKlinik/dokter/DokterService.java
@@ -6,17 +6,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import Kelompok2_RPL.AplikasiKlinik.pendaftaran.PendaftaranKonsultasi;
+import Kelompok2_RPL.AplikasiKlinik.pendaftaran.PendaftaranKonsultasiRepository;
 
 @Service
 public class DokterService {
     @Autowired
-    private DokterRepository dokterRepository;
+    private PendaftaranKonsultasiRepository pendaftaranKonsultasiRepository;
 
     public List<PendaftaranKonsultasi> getAllPendaftaran(int id_Dokter){
-        return this.dokterRepository.getAllPendaftaran(id_Dokter);
+        return this.pendaftaranKonsultasiRepository.getAllPendaftaran(id_Dokter);
     }
 
     public List<PendaftaranKonsultasi> searchPendaftaranByName(int id_Dokter, String filter){
-        return this.dokterRepository.searchPendaftaranByName(id_Dokter, filter);
+        return this.pendaftaranKonsultasiRepository.searchPendaftaranByName(id_Dokter, filter);
     }
 }
