+++ src/main/java/Kelompok2_RPL/AplikasiKlinik/pendaftaran/PendaftaranKonsultasiRepository.java
@@ -1,0 +1,8 @@
+package Kelompok2_RPL.AplikasiKlinik.pendaftaran;
+
+import java.util.List;
+
+public interface PendaftaranKonsultasiRepository {
+    List<PendaftaranKonsultasi> getAllPendaftaran(int id_Dokter);
+    List<PendaftaranKonsultasi> searchPendaftaranByName(int id_Dokter, String filter);
+}
